+++ plugin-packs/postcss-preset-env/src/lib/get-options-for-browsers-by-feature.mjs
@@ -1,9 +1,9 @@
 import browserslist from 'browserslist';
-import getUnsupportedBrowsersByFeature from './get-unsupported-browsers-by-feature';
-import { log } from './log-helper';
+import getUnsupportedBrowsersByFeature from './get-unsupported-browsers-by-feature.mjs';
+import { log } from '../log/helper.mjs';
 
 // add extra options for certain browsers by feature
-export default function getOptionsForBrowsersByFeature(browsers, feature, cssdb) {
+export default function getOptionsForBrowsersByFeature(browsers, feature, cssdbList) {
 	const supportedBrowsers = browserslist(browsers, { ignoreUnknownVersions: true });
 
 	switch (feature.id) {
@@ -18,7 +18,7 @@ export default function getOptionsForBrowsersByFeature(browsers, feature, cssdb)
 				// Nesting rules can transform selectors to use :is pseudo.
 				// This is more spec compliant but it's not supported by all browsers.
 				// If we can't use :is pseudo according to preset-env options, we add an extra option to avoid :is pseudo.
-				const feature = cssdb.find(feature => feature.id === 'is-pseudo-class');
+				const feature = cssdbList.find(feature => feature.id === 'is-pseudo-class');
 
 				if (needsOptionFor(feature, supportedBrowsers)) {
 					log('Disabling :is on "nesting-rules" due to lack of browser support.');